fix: update ignore package, rearrange for imports
