+++ core/commands/slash/built-in-legacy/onboard.ts
@@ -1,11 +1,8 @@
 import ignore from "ignore";
 
 import type { FileType, IDE, SlashCommand } from "../../..";
-import {
-  DEFAULT_IGNORE,
-  getGlobalContinueIgArray,
-  gitIgArrayFromFile,
-} from "../../../indexing/ignore";
+import { getGlobalContinueIgArray } from "../../../indexing/continueignore";
+import { DEFAULT_IGNORE, gitIgArrayFromFile } from "../../../indexing/ignore";
 import { renderChatMessage } from "../../../util/messageContent";
 import {
   findUriInDirs,
