+++ core/indexing/walkDir.ts
@@ -3,11 +3,8 @@ import ignore, { Ignore } from "ignore";
 import type { FileType, IDE } from "..";
 
 import { joinPathsToUri } from "../util/uri";
-import {
-  defaultIgnoreFileAndDir,
-  getGlobalContinueIgArray,
-  gitIgArrayFromFile,
-} from "./ignore";
+import { getGlobalContinueIgArray } from "./continueignore";
+import { defaultIgnoreFileAndDir, gitIgArrayFromFile } from "./ignore";
 
 export interface WalkerOptions {
   include?: "dirs" | "files" | "both";
